- Fixed: added some files into the solution - Added: examples for communicating with a web server through php files
