+++ GameServerScripts/web/OnlineStatusUpdateServerFiles/serverinfo.php
@@ -1,0 +1,9 @@
+<?
+
+$total = $_GET[totalclients];
+
+$fp = fopen('players.txt', 'w');
+fwrite($fp, $total);
+
+
+?>
