+++ GameServerScripts/web/OnlineStatusUpdateServerFiles/showplayers.php
@@ -1,0 +1,8 @@
+<?
+
+$filename = "players.txt";
+$handle = fopen($filename, "r");
+$contents = fread($handle, filesize($filename));
+fclose($handle);
+echo $contents;
+?>
